bug fixed

+++ src/loss.py
@@ -16,7 +16,7 @@ class ContrastiveLoss(torch.nn.Module):
         euclidean_distance = F.pairwise_distance(output1, output2)
         # loss_contrastive = torch.mean((1-label) * torch.pow(euclidean_distance, 2) +
         #                               (label) * torch.pow(torch.clamp(self.margin - euclidean_distance, min=0.0), 2))
-
+        label = label.float()
         loss_contrastive = .5 * (1 - label) * torch.pow(euclidean_distance, 2) + \
                            .5 * label * torch.pow(torch.clamp(self.margin - euclidean_distance, min=.0), 2)
